requirements.txt
pyyaml
argoverse-api
loguru
pickle
zlib
scipy
numpy
pytorch